requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0.1
tomli-w>=1.0.0